+++ classes/parte-1-responder/011-levantar-un-servidor-y-responder/implementaciones/aspnet-core/Program.cs
@@ -1,0 +1,6 @@
+var constructor = WebApplication.CreateBuilder(args);
+var app = constructor.Build();
+
+app.MapGet("/", () => Results.Text("hola", "text/plain"));
+
+app.Run();